Print absolute addresses for auli/jalr pair
It turns out that this instruction sequence can be generated to call
an absolute address.  This adds a comment to the JALR with the full
expanded address, which is useful for debugging (it turns out it's
easier to just let my PC do addition rather than having to do it in my
head).

Note that amd64 does this exact thing, so it's probably correct... :)

Signed-off-by: Palmer Dabbelt <palmer.dabbelt@eecs.berkeley.edu>