binutils, glibc: rewrite bfd; employ sane ELF ABI
The new BFD port is primarily based upon TILE-Gx, which is conveniently
simple and vanilla.  We no longer have dynamic REL relocations, and the
GOT no longer needs to be sorted a la MIPS.  So GNU-style shared object
hashes are now supported.

Note that there are probably a few lingering issues.