gcc: remove multilibs for -mfdiv/-mno-fdiv
There are a few reasons for this.  First, the ABI requires that these
instructions be emulated if not implemented in HW.  Second, the userspace
emulation library wasn't markedly faster than the kernel's.  Third, the
multilib explosion was becoming intractible.

Ultimately, the plan is to replace illegal instructions with jumps to
userspace emulation routines, which will achieve reasonable performance
portably (without multilibs).  This support is forthcoming.