binutils: rework jump pseudoisntruction
"jump" was renamed to "tail" (as in tail call) and may freely clobber a
t-register, as before.  A new "jump" was added that requires a second
argument, which is the scratch register it may clobber.

"tail" should be used for tail calls; "jump" should be used for
intraprocedural jumps more than 1 MiB away.