glibc: Undefine ELF_MACHINE_DEBUG_SETUP
Ensure that ld.so follows the preferred canonical behavior, whereby
dl_main() populates the DT_DEBUG entry in the .dynamic section with the
runtime address of the r_debug struct.  GDB relies on r_debug to locate
the link map when reconstructing the shared library list.

This is documented in elf/rtld-debugger-interface.txt in the glibc
source tree.