Port libsanitizer to RISC-V (again)
So it turns out that the uapi stuff also broke libsanitizer in some
trivial ways -- essentially we need to replace the old version of the
*at syscalls with the new *at version and some flags.  Upstream has
already fixed this, so all I really did is backport part of an
upstream patch that I don't care to track down the exact origin of:

  commit 7d752f28b590bbad13c877c2aa7f5f8de2cdff10
  Author: kcc <kcc@138bc75d-0d04-0410-961f-82ee72b054a4>
  Date:   Thu May 22 07:09:21 2014 +0000

      libsanitizer merge from upstream r209283

      git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@210743 138bc75d-0d04-0410-961f-82ee72b054a4

Note that this will:

* Break on anything that doesn't support uapi.
* Conflict with the patch that I assume will land in GCC-5.0.

Neither of which I think are important -- just don't build
riscv-gnu-toolchain for anything that's not RISC-V, and drop the patch
when you rebase to 5.0.

This is necessary to get Gentoo to build, as it builds libsanitizer.