Install the glibc and linux headers before stage1
I thought this was a problem with the build, but it turns out that was
actually related to my bad (and now reverted) patch from earlier this
week.

Regardless, this is the canonical way to build a cross compiler so now
that it's done I don't see a reason not to use it.