binutils: Add RISC-V files to ALL_MACHINES{,_CFILES}
I've been getting sporadic binutils build failures for a while that
all look something like

  .../src/binutils/bfd/cpu-riscv.c:51:5: error: ‘bfd_arch_riscv’ undeclared here (not in a function)
       bfd_arch_riscv,     \
       ^
  ...
  Makefile:1605: recipe for target 'cpu-riscv.lo' failed
  make[6]: *** [cpu-riscv.lo] Error 1
  make[6]: *** Waiting for unfinished jobs....

I belive the problem here is that "bfd.h" is in the middle of being
generated while this build step is being run.  This can happen because
of a missing dependency, which you can verify by running

  $ cd build-binutils-linux/bfd
  $ rm stmp-bfd-h bfd.h cpu-riscv.lo
  $ make cpu-riscv.lo

which will fail because "bfd.h" doesn't get regenerated.  This patch
fixes that dependency problem by adding some RISC-V files to
ALL_MACHINES{,_CFILES}, which appears to be necessary to get the
dependency to work correctly (though I have no idea where it's coming
from).

It remains to be seen if this fixes the parellal build problem, but
this is certianly a bug.