gcc: Fix the libsanitizer port, again
I guess it wouldn't be a proper GCC update if it didn't break the
libsanitizer port :).  I feel like they've picked the most convoluted
set of #ifdefs possible in order to achieve this, but I just stuck
__riscv__ in some slightly different places than last time so at least
it wasn't any work to fix.

This doesn't break riscv64-unknown-linux-gnu on a5, and fixes Gentoo.