binutils: gas: fix default xlen selection
gas currently defaults to xlen=64 all the time even when it's configured
for a riscv32-elf target.  Fix this by leveraging the existing default
arch logic like x86/s390 targets do.

The meaning of "arch" is a bit confusing in this case as we're using it
only for the xlen and not for the various extensions.  But since this is
really the only thing we pull from the tuple, and it's the naming that
the common gas code uses, oh well.