Disable newlib's dynamic rounding without F
GCC's softfloat is built by default without support for dynamic
rounding modes.  This makes softfloat newlib's floating point always
round-to-zero, and returns a failure code when trying to set it.

This patch is needed to make newlib compile on anything without F, as
it used to try use an undefined instruction to do this.