Make gas's march string case insensitive when possible
There's a comment at the beginning of this function that seems to
indicate that the ISA string checker is case insensitive, but the
implementation isn't.  This still isn't fully case insensitive: if
there's any X instructions then things can get ambiguous, so it'll
reject things like rv64iXhwacha just to be safe.