Avoid t0 for indirect calls
We sometimes use t0 as a link register, e.g. for the compressed prologues
and epilogues, and so we want microarchitectures to push the RAS when
rd=t0 and pop the RAS when rs1=t0.  Thus, we want to avoid using t0 for
indirect calls to avoid errant RAS operations.