Replace riscv/fpu/bits/mathinline.h with Tilera's version
As per but #73, this was causing some compile failures related to
nothrow when using C++ and pedantic.  The actual fix is to wrap this
with "__NTH", but Tilera's version looks simpler because it just calls
a GCC builtin rather than providing a second implementation.

I tried removing this file, but then I don't get inlined versions of
these calls.