RISC-V's mathinline.h doesn't need FPU support
This removes a difference between RISC-V's hardfloat and softfloat
implementations.  Since there's no actual code here (just a shim to
GCC's builtins), I don't think there's any reason to have this be
hardfloat specific.

I moved this because it's not in Tilera's FPU directory, and that's
where this file came from.