Check for CC and CXX at configure-time
Bug #81 resulted in a failure that was far from the actual problem
because CC was being set to "gcc-4.8" in order to try and use a new
host toolchain, which resulted in riscv-glibc being built with the
host toolchain.  In order to make these issues more apparent, this
simply errors out when CC or CXX is set.