Use "cp -a" instead of "cp -R"
"cp -a" preserves everything about the copied files, the most
important of which for us is the modification times.  It appears all
the GNU stuff has sufficiently broken toolchains such that randomizing
the order of the modification times of the distributed sources is
sufficient to break the build.  I'm hoping this fixes an intermittent
failure in my builds.