Handle lower-case ISA strings
We'd changed the ISA string parser to be case-insensitive, but when doing that
it made all "X*" extensions upper case.  The result was that the instruction
table matcher would skip "Xcustom" extensions, since they showed up as
"XCUSTOM".

This patch changes the ISA subset matching function to be case insensitive, to
match the parser.