Bump binutils to 2.25.1
This just requires a single extra argument to a function call, which I copied
from aarch64.  I've only tested that newlib builds, but I don't see how this
could cause any problems... :)

I was hoping this would help with my attempts to integrate the GDB and binutils
ports into one repo for upstreaming, but it looks like 2.25.1 is still a bit
too old for this.