Correct the "{l,s}{b,h,w,d} REG, ADDR, REG" format
ADDR in this format is only meant to be a symbolic address, not a numeric
address.  Things are this way because of linker relaxation.  Passing a numeric
address would result in an internal assertion triggering, this patch just makes
gas reject these instructions cleanly.

This compiles glibc on a5 for me.  This should fix bug #106.