gcc: "Fix" RV32D integer <> fp moves (#98)
We have no pattern for moving between a double-precision floating-point
register and an integer register in RV32.  This makes the compiler choke.

This workaround raises the cost of those moves significantly.  This has
no effect on -O0, which always worked; the other -Ox all seem to work
now.  For obvious reasons, this makes me very uncomfortable.  But,
as far as I can tell, MIPS is relying on the same behavior.

If this bug rears its head again, no big deal: we can add an "instruction"
for this, which does something like

  addi sp, sp, -16
  fsd  f4, 0(sp)
  lw   x6, 0(sp)
  lw   x7, 4(sp)
  addi sp, sp, 16

This is heinous, but it's always correct.  If the cost model is right,
the compiler will only very rarely invoke it, so the performance cost
of this ingorious hack will be hidden.  @palmer-dabbelt