gcc: make LRA the only register allocator
LRA is the new register allocator, replacing the complicated reload pass.
It's a much better design, and it seems likely that it will ultimately
replace reload altogether.  Rather than contending with the vagaries of
both schemes, I think we should abandon reload altogether.

For RISC-V, LRA seems to have a mixed performance impact, but it will
get better over time.