Start tracking binutils as a submodule
As we prepare binutils for upstreaming we want to start closely tracking
upstream changes.  There's a riscv-binutils-gdb repository that contains
the binutils port, and I want to centralize all the development there.
This commit uses that repository instead of the patches that used to
live here, so we're less likely to lose patches.