Show some make output on Travis failures
It's difficult to debug on Travis when there's no information in the
log.  We can't print the whole log because it exceeds Travis's maximum
log size, so this instead shows the output of a second "make".  Assuming
the makefiles are sane, this will produce just the failing output, but
there's an "exit 1" in there just to be sure we don't hide any parallel
make bugs.