Split the binutils and GDB builds
Binutils and GDB are kept in the same repository but follow seperate
relaese streams.  We've been release GDB versions based on binutils
tags, but this is the wrong thing to do and has caused multiple
headaches related to our out-of-tree GDB port.

This patch splits the binutils and GDB submodules, but doesn't actually
change the versions.