Support Python Virtual Environments while testing
For environment which Python 3 environment is managed by a package
manager, pip might raise an error saying "externally-managed-environment"
while testing.

To avoid this error, Python Virtual Environments (`venv`) is effective
but riscv-gnu-toolchain tries to use `pip install --user`, which, by
default, fails because "user" site-packages are not loaded under a
virtual environment.

Many recommends changing `include-system-site-packages` in
`$VENV/pyvenv.cfg` to `true` manually to avoid this kind of error but it
has a simple major pitfall: it taints the user site-packages (and ruins the
intent to make a virtual, isolated environment).

This commit tries to solve this issue by directly supporting `venv` from
riscv-gnu-toolchain.  This commit stops appending the `--user` option if
Python is running under a virtual environment.