Generate configuration files for clang/clang++
For GCC, the default `-march`/`-mabi`/`-mtune` can be specified
via configuring with `--with-arch`, `--with-abi` and `--with-tune`.
But these options won't take affect for Clang/LLVM.

For Clang/LLVM, we can use configuration files[1] to specify default
options for different tools/targets. So we automatically generate
two configuration files (for `clang` and `clang++`) and specify
these options in them.

References:
[1] https://clang.llvm.org/docs/UsersManual.html#configuration-files