CI/CD: Drop Ubuntu 20.04
We have build issues in Ubuntu 20.04 with QEMU, because of an outdated
host dependency (glib2 is 2.64.6 but >= 2.66 is required).

Ubuntu LTS releases have a standard support lifecycle of 5 years,
so Ubuntu 20.04 support will be ending in about 6 months.

Let's drop Ubuntu 20.04 support now, so we don't need a workaround
for the QEMU issue.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>