CI/CD: Bump upload-artifact to v4
We recently updated the download-artifact action to v4, because earlier
versions were deprecated and stopped working.

Now we ran into the situation, that the release pipeline broke, because
no toolchains are found by the download-artifact action
("Found 0 artifact(s)" although all toolchain builds succeeded and the
resulting toolchain was successfully archived and uploaded).

Looking into the documention of download-artifact@v4 shows the issue:
One of the breaking changes of download-artifact@v4 is "Downloading
artifacts that were created from action/upload-artifact@v3 and below
are not supported." ([1]). Our pipeline uses upload-artifact@v3, which
is why they are not found later on.

Let's bump the upload-artifact action to v4 so that download-artifact@v4
will find the built artifacts again.

[1] https://github.com/actions/download-artifact?tab=readme-ov-file#v4---whats-new

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>