Move cleanup of base image to a script and improve it
There are lots of things we can remove from the base image but
the list becomes long and it polutes the workflow file, so it
makes more sense to do this on a script instead. This gives
us up to 50GB of free space, and we don't have to wait for it to
finish, we can let those rm commands run in the background and
move forward with the rest of the build process.

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>