Use /mnt/riscv instead of /opt/riscv
We have 14GB on /mnt available (more in practice) that we don't use at all.
Instead of installing the toolchain / downloading artifacts in /opt thats
in root partition, use /mnt instead. This patch also assigns ownership
of that folder to the runner user, so that we drop use of sudo for make
later on.

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>