Reduce output size using file deduplication
There are files in the output that are the same with e.g. different names on
different paths etc (and are not symlinks). Switch them to hardlinks so that tar
doesn't archive them multiple times. This could be usefull to users too, since they
may chose to preserve hardlinks when unpacking the tarball.

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>