Allow submodule cache to persist across runs
Although it only takes a few minutes to do the shallow clones, and we
don't trigger runs that frequently, let's keep the submodule cache
around, and invalidate it based on the hash of the 'git submodules'
command, instead on each run. This will save us space on the repo's
cache storage (10GB) and speed up the process a bit.

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>