setup-apt: Add libzstd-dev to build deps
GCC and binutils will pick this up and allow compression of the debug
sections in elf files using --compress-debug-sections=zstd

The resulting binaries will depend on libzstd1.so, so users must ensure
this is installed. It seems to come by default from Ubuntu 20.04
onwards, so that shouldn't be a problem for most.