Add --disable-strip and enable gcc strip
GCC has an install target 'install-strip' that strips debug symbols at
install time. This reduces the installed size from ~4GB to ~2GB.

Enable strip on install by default but provide the option to disable it.

The downside of this is when GCC ICEs, we don't get a proper backtrace
any more.

Signed-off-by: Joel Stanley <jms@tenstorrent.com>