gdb: update submodule for gcc 15.x compatibility
Update the commit ID for the gdb submodule so that we pull a version of
gdb with the following commit from master:

|5c87b330e91 readline/tcap.h: Update definitions for C23

Without which, we see errors like the following during build (and tools
are not successfully built) if the system has gcc 15.x or newer:

|/home/tgamblin/workspace/git/riscv-gnu-toolchain/gdb/readline/readline/tcap.h:50:12: note: declared here
|   50 | extern int tgetflag ();
|      |            ^~~~~~~~
|/home/tgamblin/workspace/git/riscv-gnu-toolchain/gdb/readline/readline/terminal.c:204:26: error: too many arguments to function ‘tgetflag’; expected 0, have 1
|  204 | #define TGETFLAG(cap)   (tgetflag (cap) == TGETFLAG_SUCCESS)
|      |                          ^~~~~~~~

This occurs because of the change to gnu23 as the default C standard.
See: https://gcc.gnu.org/gcc-15/porting_to.html#c23 and the subsequent
section for details.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>