Temporary bump QEMU to latest 10.1 staging commit
Because the location of some QEMU submodules changed, we currently don't
have a stable QEMU release that can be built.
The issue is fixed on master and in the staging 10.1 branch.
To unbreak the build, let's temporary bump QEMU to the latest commit
on the staging-10.1 branch.

Of course, as soon as a new QEMU release is available, we should
redirect the submodule to it.

Fixes #1771.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>